+++ pkg/apis/extensions/v1alpha1/helper/operatingsystemconfig.go
