+++ pkg/apis/extensions/v1alpha1/helper/worker.go
