+++ pkg/apis/operator/v1alpha1/helper/garden.go
